Move lexer to new folder
