+++ server/src/common.ts
@@ -4,7 +4,7 @@ import { readdirSync, readFileSync, existsSync, statSync, promises as fs } from 
 import { Connection, MessageConnection } from 'vscode-languageserver';
 import { TextDocument } from 'vscode-languageserver-textdocument';
 import { CompletionItem, CompletionItemKind, Hover, InsertTextFormat, Range, SymbolKind } from 'vscode-languageserver-types';
-import { AhkSymbol, Lexer, fixupFormatConfig, updateCommentTagRegex } from './lexer';
+import { AhkSymbol, Lexer, fixupFormatConfig, updateCommentTagRegex } from './lexer/lexer';
 import { diagnostic } from './localize';
 import { jsDocTagNames } from './constants';
 import { AHKLSConfig, CfgKey, getCfg, LibIncludeType, setCfg, setConfigRoot } from '../../util/src/config';
@@ -17,7 +17,7 @@ export * from './constants';
 export * from './definitionProvider';
 export * from './formattingProvider';
 export * from './hoverProvider';
-export * from './lexer';
+export * from './lexer/lexer';
 export * from './localize';
 export * from './referencesProvider';
 export * from './renameProvider';
