+++ server/src/lexer/lexer.ts
@@ -18,15 +18,15 @@ import {
 
 import { TextDocument } from 'vscode-languageserver-textdocument';
 import { URI } from 'vscode-uri';
-import { builtin_ahkv1_commands, builtin_variable, builtin_variable_h } from './constants';
-import { action, completionitem, diagnostic, warn } from './localize';
+import { builtin_ahkv1_commands, builtin_variable, builtin_variable_h } from '../constants';
+import { action, completionitem, diagnostic, warn } from '../localize';
 import {
 	a_vars, ahk_version, ahkuris, ahkvars, alpha_3, connection,
 	hoverCache, isahk2_h, lexers, libdirs, libfuncs, locale, openAndParse, openFile,
 	restorePath, rootdir, setTextDocumentLanguage, symbolProvider, utils, workspaceFolders
-} from './common';
-import { ActionType, BlockStyle, BraceStyle, CallWithoutParentheses, CfgKey, FormatOptions, getCfg } from '../../util/src/config';
-import { shouldExclude } from '../../util/src/exclude';
+} from '../common';
+import { ActionType, BlockStyle, BraceStyle, CallWithoutParentheses, CfgKey, FormatOptions, getCfg } from '../../../util/src/config';
+import { shouldExclude } from '../../../util/src/exclude';
 
 export interface ParamInfo {
 	offset: number
